raphael: motor: Add a warning dialog for manual calibrations
* This is a combination of the following 3 commits:

raphael: motor: Add a warning dialog for manual calibrations

Change-Id: Id764e1cbe1a8c535c0039938097465d9a0dcd02b

raphael: motor: Always show the warning dialog

Change-Id: I979cc479ee928edf0e203489264df340326fbf5b

raphael: motor: Remove MotorCalibrationWarningDialog class

Instead of using a class for the AlertDialog just create and show it directly.

This is done in preparation for upgrading the sdk. Fragments must be a public
static class to be properly recreated from instance state.

Change-Id: I32cd2e9310ebadc2c9c5da0afbac015dd32ce0a6